requirements.txt
yfinance
pandas
numpy
rich
plyer
pyyaml
pytz
requests
